requirements.txt
# Interface
streamlit>=1.32.0
plotly>=5.19.0

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# Download e leitura de dados
requests>=2.31.0
pyarrow>=15.0.0

# Pipeline de dados
duckdb>=0.10.0
supabase>=2.4.0

# Experimentos
mlflow>=2.11.0

# Utilitários
python-dotenv>=1.0.0